adds test definitions for constructor integrations
