+++ lib/db/database.js
@@ -71,6 +71,7 @@ class Database extends RegentObject {
 
         const connections = this.call(getConnections);
         $protected.set(this, { connections });
+        this.call(setConnectionEvents);
     }
 
     /**
@@ -212,6 +213,15 @@ function getConnections() {
         read,
         write,
     };
+}
+
+/**
+ * Initialize the connections on an object
+ *
+ * @method setConnectionEvents
+ */
+function setConnectionEvents() {
+    const { connections } = $protected(this);
 }
 
 module.exports = Database;
